Implement regex mappers
This is an experimental implementation for regex mappers.

The use case is the following: One wants to record data from, for
example, `free`. The command reports data in the following format:

                  total        used        free      shared  buff/cache   available
    Mem:       16396624     2593408      528324      107556    13274892    13357476
    Swap:             0           0           0

Now, one wants to save this data in a "more structured way" than plain
text.

The feature from this patch allows the user to create regexes which
match certain text elements. These text elements are then stored
one-by-one in the "database".

The database format for these entries is as follows:

old format:

    {timestamp} {command output}

the new format (only for rows which use the regex mapper, else old
format):

    {timestamp} {mapper}.{match} {output}

Explanation:

* The "mapper" is (in this case the regex mapper) a simple numeric
  value which represents the index of the mapper. So the first mapper gets
  `0`, the next one `1` and so on.
* The "match" is the number of the match (the index of the output
  value). As regexes can match multiple groups (think of `(.*)-(.*)`), a
  regex can "report" multiple values back.

Of course:

    number of database entries =
        if no mappers:  1
        else:           #mappers * #outputs_per_mapper

Theoretically, the crate could support built-in mappers with this, for
example "extract the first number from the output" or even "run the
output JSON through this `jq` command before storing.

The implementation is not completed yet, as the whole implementation of
mappers depends on regex right now, meaning that adding other mappers is
not trivial right now.

Either way, it is a POC.

Suggested-by: Mario Krehl <mario-krehl@gmx.de>
Signed-off-by: Matthias Beyer <mail@beyermatthias.de>